+++ src/core/thread/link_metrics.hpp
@@ -70,152 +70,6 @@ namespace LinkMetrics {
  */
 
 /**
- * This type represents the results (values) for a set of metrics.
- *
- * @sa otLinkMetricsValues.
- *
- */
-class MetricsValues : public otLinkMetricsValues, public Clearable<MetricsValues>
-{
-public:
-    /**
-     * This method gets the metrics flags.
-     *
-     * @returns The metrics flags.
-     *
-     */
-    Metrics &GetMetrics(void) { return static_cast<Metrics &>(mMetrics); }
-
-    /**
-     * This method gets the metrics flags.
-     *
-     * @returns The metrics flags.
-     *
-     */
-    const Metrics &GetMetrics(void) const { return static_cast<const Metrics &>(mMetrics); }
-
-    /**
-     * This method set the metrics flags.
-     *
-     * @param[in] aMetrics  The metrics flags to set from.
-     *
-     */
-    void SetMetrics(const Metrics &aMetrics) { mMetrics = aMetrics; }
-};
-
-/**
- * This class represents one Series that is being tracked by the Subject.
- *
- * When an Initiator successfully configured a Forward Tracking Series, the Subject would use an instance of this class
- * to track the information of the Series. The Subject has a `Pool` of `SeriesInfo`. It would allocate one when a new
- * Series comes, and free it when a Series finishes.
- *
- * This class inherits `LinkedListEntry` and each `Neighbor` has a list of `SeriesInfo` so that the Subject could track
- * per Series initiated by neighbors as long as it has available resources.
- *
- */
-class SeriesInfo : public LinkedListEntry<SeriesInfo>
-{
-    friend class LinkedList<SeriesInfo>;
-    friend class LinkedListEntry<SeriesInfo>;
-
-public:
-    /**
-     * This constant represents Link Probe when filtering frames to be accounted using Series Flag. There's
-     * already `kFcfFrameData`, `kFcfFrameAck` and `kFcfFrameMacCmd`. This item is added so that we can
-     * filter a Link Probe for series in the same way as other frames.
-     *
-     */
-    static constexpr uint8_t kSeriesTypeLinkProbe = 0;
-
-    /**
-     * This method initializes the SeriesInfo object.
-     *
-     * @param[in]  aSeriesId      The Series ID.
-     * @param[in]  aSeriesFlags   The Series Flags which specify what types of frames are to be accounted.
-     * @param[in]  aMetrics       Metrics to query.
-     *
-     */
-    void Init(uint8_t aSeriesId, const SeriesFlags &aSeriesFlags, const Metrics &aMetrics);
-
-    /**
-     * This method gets the Series ID.
-     *
-     * @returns  The Series ID.
-     *
-     */
-    uint8_t GetSeriesId(void) const { return mSeriesId; }
-
-    /**
-     * This method gets the PDU count.
-     *
-     * @returns  The PDU count.
-     *
-     */
-    uint32_t GetPduCount(void) const { return mPduCount; }
-
-    /**
-     * This method gets the average LQI.
-     *
-     * @returns  The average LQI.
-     *
-     */
-    uint8_t GetAverageLqi(void) const { return mLqiAverager.GetAverage(); }
-
-    /**
-     * This method gets the average RSS.
-     *
-     * @returns  The average RSS.
-     *
-     */
-    int8_t GetAverageRss(void) const { return mRssAverager.GetAverage(); }
-
-    /**
-     * This method aggregates the Link Metrics data of a frame into this series.
-     *
-     * @param[in]  aFrameType    The type of the frame.
-     * @param[in]  aLqi          The LQI value.
-     * @param[in]  aRss          The RSS value.
-     *
-     */
-    void AggregateLinkMetrics(uint8_t aFrameType, uint8_t aLqi, int8_t aRss);
-
-    /**
-     * This methods gets the metrics.
-     *
-     * @returns  The metrics associated with `SeriesInfo`.
-     *
-     */
-    const Metrics &GetLinkMetrics(void) const { return mMetrics; }
-
-private:
-    bool Matches(const uint8_t &aSeriesId) const { return mSeriesId == aSeriesId; }
-    bool IsFrameTypeMatch(uint8_t aFrameType) const;
-
-    SeriesInfo *mNext;
-    uint8_t     mSeriesId;
-    SeriesFlags mSeriesFlags;
-    Metrics     mMetrics;
-    RssAverager mRssAverager;
-    LqiAverager mLqiAverager;
-    uint32_t    mPduCount;
-};
-
-/**
- * This enumeration type represent Link Metrics Status.
- *
- */
-enum Status : uint8_t
-{
-    kStatusSuccess                   = OT_LINK_METRICS_STATUS_SUCCESS,
-    kStatusCannotSupportNewSeries    = OT_LINK_METRICS_STATUS_CANNOT_SUPPORT_NEW_SERIES,
-    kStatusSeriesIdAlreadyRegistered = OT_LINK_METRICS_STATUS_SERIESID_ALREADY_REGISTERED,
-    kStatusSeriesIdNotRecognized     = OT_LINK_METRICS_STATUS_SERIESID_NOT_RECOGNIZED,
-    kStatusNoMatchingFramesReceived  = OT_LINK_METRICS_STATUS_NO_MATCHING_FRAMES_RECEIVED,
-    kStatusOtherError                = OT_LINK_METRICS_STATUS_OTHER_ERROR,
-};
-
-/**
  * This class implements Thread Link Metrics query and management.
  *
  */
@@ -454,11 +308,6 @@ private:
  */
 
 } // namespace LinkMetrics
-
-DefineCoreType(otLinkMetrics, LinkMetrics::Metrics);
-DefineCoreType(otLinkMetricsValues, LinkMetrics::MetricsValues);
-DefineMapEnum(otLinkMetricsEnhAckFlags, LinkMetrics::EnhAckFlags);
-
 } // namespace ot
 
 #endif // OPENTHREAD_CONFIG_MLE_LINK_METRICS_INITIATOR_ENABLE || OPENTHREAD_CONFIG_MLE_LINK_METRICS_SUBJECT_ENABLE
